Add parse_to_url, TryFrom and GitProvider for Url
